+++ client/src/app/components/rdio-scanner/rdio-scanner.ts
@@ -1,6 +1,6 @@
 /*
  * *****************************************************************************
- * Copyright (C) 2019-2020 Chrystian Huot
+ * Copyright (C) 2019-2021 Chrystian Huot
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -85,6 +85,7 @@ export interface RdioScannerEvent {
     groups?: RdioScannerGroup[];
     holdSys?: boolean;
     holdTg?: boolean;
+    linked?: boolean;
     livefeedMode?: RdioScannerLivefeedMode;
     map?: RdioScannerLivefeedMap;
     pause?: boolean;